+++ wled00/FX.h
@@ -584,7 +584,7 @@ class WS2812FX {
     }
 
     void
-      finalizeInit(bool supportWhite, uint16_t countPixels, bool skipFirst),
+      finalizeInit(uint16_t countPixels, bool skipFirst),
       service(void),
       blur(uint8_t),
       fill(uint32_t),
@@ -608,6 +608,7 @@ class WS2812FX {
       setPixelSegment(uint8_t n);
 
     bool
+      isRgbw = false,
       gammaCorrectBri = false,
       gammaCorrectCol = true,
       applyToAllSelected = true,
@@ -813,7 +814,6 @@ class WS2812FX {
     void handle_palette(void);
 
     bool
-      _useRgbw = false,
       _skipFirstMode,
       _triggered;