+++ wled00/ir.cpp
@@ -345,19 +345,19 @@ void decodeIR40(uint32_t code)
     case IR40_MAGENTA      : colorFromUint24(COLOR_MAGENTA);                             break;
     case IR40_PINK         : colorFromUint24(COLOR_PINK);                                break;
     case IR40_WARMWHITE2   : {
-      if (useRGBW) {        colorFromUint32(COLOR2_WARMWHITE2);   effectCurrent = 0; }    
+      if (strip.isRgbw) {        colorFromUint32(COLOR2_WARMWHITE2);   effectCurrent = 0; }    
       else                  colorFromUint24(COLOR_WARMWHITE2);                       }   break;
     case IR40_WARMWHITE    : {
-      if (useRGBW) {        colorFromUint32(COLOR2_WARMWHITE);    effectCurrent = 0; }    
+      if (strip.isRgbw) {        colorFromUint32(COLOR2_WARMWHITE);    effectCurrent = 0; }    
       else                  colorFromUint24(COLOR_WARMWHITE);                        }   break;
     case IR40_WHITE        : {
-      if (useRGBW) {        colorFromUint32(COLOR2_NEUTRALWHITE); effectCurrent = 0; }    
+      if (strip.isRgbw) {        colorFromUint32(COLOR2_NEUTRALWHITE); effectCurrent = 0; }    
       else                  colorFromUint24(COLOR_NEUTRALWHITE);                     }   break;
     case IR40_COLDWHITE    : {
-      if (useRGBW) {        colorFromUint32(COLOR2_COLDWHITE);    effectCurrent = 0; }   
+      if (strip.isRgbw) {        colorFromUint32(COLOR2_COLDWHITE);    effectCurrent = 0; }   
       else                  colorFromUint24(COLOR_COLDWHITE);                        }   break;
     case IR40_COLDWHITE2    : {
-      if (useRGBW) {        colorFromUint32(COLOR2_COLDWHITE2);   effectCurrent = 0; }   
+      if (strip.isRgbw) {        colorFromUint32(COLOR2_COLDWHITE2);   effectCurrent = 0; }   
       else                  colorFromUint24(COLOR_COLDWHITE2);                       }   break;
     case IR40_WPLUS        : relativeChangeWhite(10);                                    break;
     case IR40_WMINUS       : relativeChangeWhite(-10, 5);                                break;
@@ -402,21 +402,21 @@ void decodeIR44(uint32_t code)
     case IR44_MAGENTA     : colorFromUint24(COLOR_MAGENTA);                             break;
     case IR44_PINK        : colorFromUint24(COLOR_PINK);                                break;
     case IR44_WHITE       : {
-      if (useRGBW) {
+      if (strip.isRgbw) {
         if (col[3] > 0) col[3] = 0; 
         else {              colorFromUint32(COLOR2_NEUTRALWHITE); effectCurrent = 0; }
       } else                colorFromUint24(COLOR_NEUTRALWHITE);                     }  break;
     case IR44_WARMWHITE2  : {
-      if (useRGBW) {        colorFromUint32(COLOR2_WARMWHITE2);   effectCurrent = 0; }    
+      if (strip.isRgbw) {        colorFromUint32(COLOR2_WARMWHITE2);   effectCurrent = 0; }    
       else                  colorFromUint24(COLOR_WARMWHITE2);                       }  break;
     case IR44_WARMWHITE   : {
-      if (useRGBW) {        colorFromUint32(COLOR2_WARMWHITE);    effectCurrent = 0; }    
+      if (strip.isRgbw) {        colorFromUint32(COLOR2_WARMWHITE);    effectCurrent = 0; }    
       else                  colorFromUint24(COLOR_WARMWHITE);                        }  break;
     case IR44_COLDWHITE   : {
-      if (useRGBW) {        colorFromUint32(COLOR2_COLDWHITE);    effectCurrent = 0; }   
+      if (strip.isRgbw) {        colorFromUint32(COLOR2_COLDWHITE);    effectCurrent = 0; }   
       else                  colorFromUint24(COLOR_COLDWHITE);                        }  break;
     case IR44_COLDWHITE2  : {
-      if (useRGBW) {        colorFromUint32(COLOR2_COLDWHITE2);   effectCurrent = 0; }    
+      if (strip.isRgbw) {        colorFromUint32(COLOR2_COLDWHITE2);   effectCurrent = 0; }    
       else                  colorFromUint24(COLOR_COLDWHITE2);                       }  break;
     case IR44_REDPLUS     : relativeChange(&effectCurrent,  1, 0, MODE_COUNT);          break;
     case IR44_REDMINUS    : relativeChange(&effectCurrent, -1, 0);                      break;