+++ wled00/led.cpp
@@ -50,14 +50,14 @@ void setAllLeds() {
   {
     strip.setBrightness(scaledBri(briT));
   }
-  if (useRGBW && strip.rgbwMode == RGBW_MODE_LEGACY)
+  if (strip.isRgbw && strip.rgbwMode == RGBW_MODE_LEGACY)
   {
     colorRGBtoRGBW(col);
     colorRGBtoRGBW(colSec);
   }
   strip.setColor(0, col[0], col[1], col[2], col[3]);
   strip.setColor(1, colSec[0], colSec[1], colSec[2], colSec[3]);
-  if (useRGBW && strip.rgbwMode == RGBW_MODE_LEGACY)
+  if (strip.isRgbw && strip.rgbwMode == RGBW_MODE_LEGACY)
   {
     col[3] = 0; colSec[3] = 0;
   }