+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 ## WLED changelog
 
-### Development versions after 0.11.1 release
+### Development versions between 0.11.1 and 0.12.0 releases
+
+#### Build 2103290
+
+-   Version bump to 0.12.0-b4 "Hikari"
+-   Experimental use of espressif32@3.1.1
+-   Fixed RGBW mode disabled after LED settings saved
+-   Fixed infrared support not compiled in if IRPIN is not defined
 
 #### Build 2103230
 
@@ -233,7 +240,7 @@
 #### Build 2011153
 
 -   Fixed an ESP32 end-of-file issue
--   Fixed useRGBW not read from cfg.json
+-   Fixed strip.isRgbw not read from cfg.json
 
 #### Build 2011152